Moved some tests around so that they match the standards. This allows me to see that I have implemented all of a section and to use the tests from the standard more easily. Might change it eventually based on other categories.
Also, I realized that I was making an error with display vs printing the
forms. When you print one from the interpreter it is the same as what
you would read. I.e. "hello \n world" should show print that way, but
display without the quotes and have an actual newline. Gave the some
classes a to_extern method to convert to the external representation and
will also give a to_string or the Display as the way to show what it
should print as. This probably needs more work, but it is close now.