A lot of changes were needed to implement some list stuff.
Pairs can be cyclic and when cyclic or dotted fail the list? proc. To deal with
this list? has to  traverse the whole list to tell if it is a list or not, i.e. it
is finite and terminates with '(). To accomplish these things and to
make it possible to tell if a list is dotted or cyclic I adjusted the
list iterators. They now store a list of seen pointers and check for
cyclic lists. They also check for dotted lists. To make this work most
places list_to_vec uses a pair iterator and checks the cells to see if
the last one is dotted and the iterator returns undefined when the last
element is cyclic. These can be caught and used to pass some flags back.

Since we almost always convert a list to a vector when evaluating right
now it is easy to check for imporoper lists and cycles. However, now
some things are going to be o(n) when maybe they could be less. Either
way the tests all pass and that is most important for now. Though there
may need to be some adjustments to the code to clean up the issues that
were introduced by these changes.

Also, since we needed immutable and mutable pairs I got rid of dotted
pairs and added immutable pairs. These are treated properly now, but I
am not sure that will always be the case. I.e. it is a real pain in many
places to deal with the types separately since their pointer types are
different. If I can figure out how to turn an Rc<RefCell<_>> into an
Rc<_> maybe it could be fixed. But I doubt that is possible. If we can
find a way to treat them the same it will be great, but if not some care
will have to be taken when creating lists internally as only immutable
pairs can be evalled right now. Otherwise the whole switch will have to
be duplicated.