Adjusted the external representation and display for ScmVal so that it shares code where possible. The biggest change was to use list_to_vec for the list display/extern functions. Also, anywhere both are the same display uses external representation.
Fully removed the hash map types. Given the issues I ran into trying to
hash things to maintain the seen list in the cell iterators it seems
that hash tables will take some extra work that I do not want to do just
yet. To keep them from clutering the code and making more work for some
things I removed them. This includes the wrapper class. The env now just
uses a regular hash map for it's scopes. The only thing that is hashable
are symbols and they are checked in the env insert so nothing else is
used as a key. In the future if I want to add hash maps I will need to
do some work to figure out how people hash lists etc. The pointer method
that I used in the iterator might work, provided that pointer is stable
and the memory location of an Rc or RefCell never changes, but I am not
sure I can guarantee that.