Added list length, but had to make some changes to list?. The test was copied and still using null? so it was passing, but would error if list? was called on a non-pair. This was fixed.
At the same time I realized that the way I expected lists to be printed
when they were cyclic was not happening properly. I fixed this too by
adding a token and value for #cyclic#. However, it is an error for the
user to evaluate the token. It is probably a problem for external
representations, but as we do not have to support cyclic lists they
should not be expected to be printable and then readable... that is for
r7rs.