Added external representations and made cycle checking less verbose. Tested all of the new methods.
NOTE that cycle checking is not perfect. It will catch references inside
of the list to the same list, but will not work if, say you have two
lists where one refers to another as an element and the second list has
an element or a tail that refers back to the first list. Also, not sure
if vectors will need cycle checking too.