Updated cell to use references more instead of cloning things. This means the iterators now take a reference to a cell instead of an Rc. All of these changes work because everything is immutable so with a lifetime we can pass around references instead of cloning Rc while we come up with a final value. For many functions the goal is to end up with a clone of a value, but ideally that is at the end of the work being done when a new or copied value is created.
The cell functions for scheme lists were moved back out of the struct
and implemented as list procedures in a separate file. They work pretty
much as the previous ones did, but with updates for the new cell etc.
They are kind of in between the previous functions in core_proc and a
method on a cell. They take and return all values, but some return
references. The references could be used in other functions that need
list processing, but may be a waste as the caller might just clone them
anyway to be passed back as a copied value. They do check their
argument types, but currently return errors that are not very useful.
Either the errors need to change, or they still need wrappers in
core_procs. I can see them just being called in core procs, so if that
is the case then the errors will need to be updated to reflect the full
informations. However, if we wrap them in core_procs then more complex
errors can be build there from the results.