Added some files to put the other procedure types into instead of sandwiching them into one file, makes it more likely that I will test them.
Also, made some adjustments to how Proc objects work. I made a second
enum for arity to represent the same reality that the formals do for
closures. Formals are necessary for closures because they provide a name
we can use to bind the args in the environment, but that is unecessary
for procedures. However, they still need an arity and the type list
feels insufficient for this, though it may need some adaptations for
rest arguments as well. Maybe the arity can be made to hold types just
like the formals hold Str, but for now it holds numbers and makes it
easy for us to check before calling the func that the number of
arguments will be satisfied.