Made some additional helpers for the proc utils to make it easier to check that an args list is a list and reduce some code duplication.
Cleaned up the array procs using some of the above as well as adding a
new extractor to Value to cover uints and some new error types. The
error types now collect more information. The final error generally just
adds the name. We may be able to simplify this and add the name earlier,
but this works for better errors and doesn't require changing as much.

Adjusted the error messages for some of the new types and put a couple
tests in the vector test file to check that they are written as
expected. Generally errors should be tested in rust as testing them in
the scheme tests is brittle if the error messages change. Later I can
remake my error testing functions in help.rs to add them in a more
robust way.