Fixed the compiler to compile the program with it's own environment. It adds a layer of indirection to variable lookups, but given the way that rust deals with captures for closures it is not possible to compile my code to rust in a simple way (or at all). So we add the environment to be able to capture it for closures and just lookup symbols there instead of storing them in rust vars. There are only a few forms that compile, but given the use of the builtin procs and the env it is not too hard to add new forms.
I also added a set of tests that compile the program and write it to a
rust file in a test program folder. This allows compiling to rust and
compiling and running the rust program to check the correctness of the
compilation and the result of the program. It is not perfect yet as it
overwites the files and tests must all be run in the same #[test]
function, but for the moment it allows a good way to test different
compilations that might prove tricky. Also, since it writes the compiled
code to a persistent file we can check that file and make adjustments
and run it until we know what works and then make the modifications to
the compiler and try again.