Adjusted the compiler again so that the module compiler is where all of the compile methods are. It will be responsible for constructing a compilation unit for a file. If a file is the main file then it will have a main function added below it where the module function is called.
I also copied the core file to the lib as that way the compiled code
only relies on the lib and not on the compilers rust code. I am not sure
I want to call it hlep.rs when I already have util.rs, but core.rs also
sounded wrong, so it can be help for now.

The compiler lib tests work again with the new expected output, however
the tests folder tests have not been tested and likely will not work
yet. I have to fix up a Program compiler for that and confirm that
everything the files need to have in order to compile with rust is
included.