Removed some duplicates(probably not all of them). Started the program compiler. Fixed up the module compiler so that it can be tested.
Note that the symbols being instantiated before the program does not
work as I thought it might. It has the same problem as the original
closure issue where the lambda function is a closure that moves the
symbol. So the way to improve efficiency would be to figure out a new
way to do the lambda that fixes the original problem without the
environment.