Added the tests folder to the lib project. Eventually, I will move the interpreter to it's own crate that calls the lib and the tests can move there. The lib can just use unit tests for all its data and functions.
Made some minor adjustments to the compiler and added a mutual recursion
test. It shows that Rust does not guarantee tco under all circumstances.
Perhaps with full optimizations it would do it, but if it is not a
guarantee then it is a bad idea to compile scheme to rust without some
more runtime guarantees. I expect the happy medium between the compiler
and the interpreter is either a vm or some stack added to the rust
version for function applications to use. The major inefficiency of the
interpreter right now is that it saves unevaluated code that is then
evaluated all the time. If we could reduce this to rust code that could
be executed by rust, but with a little more hand holding it might do the
trick for compiled code without requiring a full interpreter. The other
option is to build it into a byte code vm so that code is compiled to
byte code before it is run and then byte code can be saved for a
functions body to reduce the issues there. There are probably other
solutions that use builtin/unsafe code to manipulate allocated memory
without a bytecode interpreter, but still with an interpreter and
runtime. I do like being able to take advantage of rust though. And I
suppose that the unevaluated code is still a rust data structure. It is
not like it has to be read in again, though the macro like evaluation
steps happen at evaluaiton time and not at read time, so something like
do is transformed everytime from do->letrec->let->lambda.